# bugfix: fix merge_state_v2 cuda graph