fix https://github.com/sgl-project/sglang/issues/5404

Bind the v2 merge_state kernel to the CUDA stream as the CUDA graph requires it.

- cmd

```bash
MERGE_STATE_MODE=cuda python3 -m sglang.launch_server \
  --model $HF_MODELS/Llama-3.1-8B-Instruct \
  --trust-remote-code \
  --tp-size 1 \
  --mem-fraction-static 0.9 \
  --attention-backend flashinfer \
  --host 0.0.0.0 --port 8852 \
  --dtype bfloat16
```
to test merge_state v2 with flashinfer, please modify flashinfer_backend.py, like:

```python
from sgl_kernel import merge_state_v2

# define merge_state_triton ......

import os

MERGE_STATE_MODE = os.environ.get("MERGE_STATE_MODE", "cuda").lower()

print(f"MERGE_STATE_MODE: {MERGE_STATE_MODE}")
if MERGE_STATE_MODE == "cuda":
    merge_state = merge_state_v2
else:
    merge_state = merge_state_triton
```


- before this fix

```bash
python3 benchmark/gsm8k/bench_sglang.py \
  --port 8852 --num-shots 8 \
  --num-questions 128 \
  --parallel 16 \
  --data-path /workspace/dev/openllm/benchmarks/data/gsm8k/test.jsonl
100%|███████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████| 128/128 [01:39<00:00,  1.29it/s]
Accuracy: 0.000
Invalid: 0.945
Latency: 99.670 s
```
Accuracy is 0.000

- after this fix

```bash
 python3 benchmark/gsm8k/bench_sglang.py \
  --port 8852 --num-shots 8 \
  --num-questions 128 \
  --parallel 16 \
  --data-path /workspace/dev/openllm/benchmarks/data/gsm8k/test.jsonl
100%|███████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████| 128/128 [00:24<00:00,  5.15it/s]
Accuracy: 0.781
Invalid: 0.000
Latency: 25.067 s
```
Accuracy is 0.781

I have only conducted tests under the Ada architecture so far. Please carry out more tests on other architectures.